plumb: add support for man,file,dir
file and dir support is based on cache stored in path referred by
fileCache and dirCache respectively in config.go

each file/dir to check against is listed with a newline as the
separator.

man pages are checked only if the string contains a (. if the str
has one, then it is checked if the entry is present in man database
using the apropos command. note: if you use mandoc, then if the
mandoc.db is not up-to-date, this will fail even if you have the
manpage in MANPATH.